backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
python-dotenv==1.0.1
Flask-MySQLdb==1.0.1
pymongo==4.6.3
dnspython==2.6.1
email-validator==2.1.1
bcrypt==4.1.2
itsdangerous==2.1.2

